# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Deltaswap Phylax API\nThis is the API for the Deltaswap Phylax and Explorer.\nThe API has two namespaces: deltaswapscan and phylax.\ndeltaswapscan is the namespace for the explorer and the new endpoints. The prefix is /api/v1.\nphylax is the legacy namespace backguard compatible with phylax node API. The prefix is /v1.\nThis API is public and does not require authentication although some endpoints are rate limited.\nCheck each endpoint documentation for more information.",
        "title": "Deltaswapscan API",
        "termsOfService": "https://wormhole.com/",
        "contact": {
            "name": "API Support",
            "url": "https://discord.com/invite/wormholecrypto",
            "email": "info@wormhole.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/governor/config": {
            "get": {
                "description": "Returns governor configuration for all phylaxs.",
                "tags": [
                    "deltaswapscan"
                ],
                "operationId": "governor-config",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number.",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of elements per page.",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response-governor_GovConfig"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/governor/config/:phylax_address": {
            "get": {
                "description": "Returns governor configuration for a given phylax.",
                "tags": [
                    "deltaswapscan"
                ],
                "operationId": "governor-config-by-phylax-address",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response-governor_GovConfig"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/governor/enqueued_vaas/": {
            "get": {
                "description": "Returns enqueued VAAs for each blockchain.",
                "tags": [
                    "deltaswapscan"
                ],
                "operationId": "governor-enqueued-vaas",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number.",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of elements per page.",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "ASC",
                            "DESC"
                        ],
                        "type": "string",
                        "description": "Sort results in ascending or descending order.",
                        "name": "sortOrder",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response-array_governor_EnqueuedVaas"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/governor/enqueued_vaas/:chain": {
            "get": {
                "description": "Returns all enqueued VAAs for a given blockchain.",
                "tags": [
                    "deltaswapscan"
                ],
                "operationId": "phylaxs-enqueued-vaas-by-chain",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number.",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of elements per page.",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "ASC",
                            "DESC"
                        ],
                        "type": "string",
                        "description": "Sort results in ascending or descending order.",
                        "name": "sortOrder",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response-array_governor_EnqueuedVaaDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/governor/limit": {
            "get": {
                "description": "Returns the governor limit for all blockchains.",
                "tags": [
                    "deltaswapscan"
                ],
                "operationId": "governor-notional-limit",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number.",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of elements per page.",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response-array_governor_GovernorLimit"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/governor/notional/available": {
            "get": {
                "description": "Returns the amount of notional value available for each blockchain.",
                "tags": [
                    "deltaswapscan"
                ],
                "operationId": "governor-notional-available",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number.",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of elements per page.",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "ASC",
                            "DESC"
                        ],
                        "type": "string",
                        "description": "Sort results in ascending or descending order.",
                        "name": "sortOrder",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response-array_governor_NotionalAvailable"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/governor/notional/available/:chain": {
            "get": {
                "description": "Returns the amount of notional value available for a given blockchain.",
                "tags": [
                    "deltaswapscan"
                ],
                "operationId": "governor-notional-available-by-chain",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number.",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of elements per page.",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response-array_governor_NotionalAvailableDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/governor/notional/limit": {
            "get": {
                "description": "Returns the detailed notional limit for all blockchains.",
                "tags": [
                    "deltaswapscan"
                ],
                "operationId": "governor-notional-limit-detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number.",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of elements per page.",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response-array_governor_NotionalLimitDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/governor/notional/limit/:chain": {
            "get": {
                "description": "Returns the detailed notional limit available for a given blockchain.",
                "tags": [
                    "deltaswapscan"
                ],
                "operationId": "governor-notional-limit-detail-by-chain",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number.",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of elements per page.",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response-array_governor_NotionalLimitDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/governor/notional/max_available/:chain": {
            "get": {
                "description": "Returns the maximum amount of notional value available for a given blockchain.",
                "tags": [
                    "deltaswapscan"
                ],
                "operationId": "governor-max-notional-available-by-chain",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response-governor_MaxNotionalAvailableRecord"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/governor/status": {
            "get": {
                "description": "Returns the governor status for all phylaxs.",
                "tags": [
                    "deltaswapscan"
                ],
                "operationId": "governor-status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number.",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of elements per page.",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response-array_governor_GovStatus"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/governor/status/:phylax_address": {
            "get": {
                "description": "Returns the governor status for a given phylax.",
                "tags": [
                    "deltaswapscan"
                ],
                "operationId": "governor-status-by-phylax-address",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number.",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of elements per page.",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response-governor_GovStatus"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/v1/governor/enqueued_vaas": {
            "get": {
                "description": "Get enqueued VAAs",
                "tags": [
                    "Phylax"
                ],
                "operationId": "phylaxs-enqueued-vaas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/governor.EnqueuedVaaResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "governor.EnqueuedVaaDetail": {
            "type": "object",
            "properties": {
                "chainId": {
                    "$ref": "#/definitions/vaa.ChainID"
                },
                "emitterAddress": {
                    "type": "string"
                },
                "notionalValue": {
                    "type": "integer"
                },
                "releaseTime": {
                    "type": "integer"
                },
                "sequence": {
                    "type": "string"
                },
                "txHash": {
                    "type": "string"
                }
            }
        },
...
        "governor.EnqueuedVaaResponse": {
            "type": "object",
            "properties": {
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/governor.EnqueuedVaaItemResponse"
                    }
                }
            }
        },
        "governor.EnqueuedVaas": {
            "type": "object",
            "properties": {
                "chainId": {
                    "$ref": "#/definitions/vaa.ChainID"
                },
                "enqueuedVaas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/governor.EnqueuedVaa"
                    }
                }
            }
        },
        "governor.GovConfig": {
            "type": "object",
            "properties": {
                "chains": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/governor.GovConfigChains"
                    }
                },
                "counter": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nodeName": {
                    "type": "string"
                },
                "tokens": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/governor.GovConfigfTokens"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "response.Response-array_governor_EnqueuedVaaDetail": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/governor.EnqueuedVaaDetail"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/response.ResponsePagination"
                }
            }
        },
        "response.Response-array_governor_EnqueuedVaas": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/governor.EnqueuedVaas"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/response.ResponsePagination"
                }
            }
        },
        "response.Response-array_governor_GovStatus": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/governor.GovStatus"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/response.ResponsePagination"
                }
            }
        },
        "response.Response-array_governor_GovernorLimit": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/governor.GovernorLimit"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/response.ResponsePagination"
                }
            }
        },
        "response.Response-array_governor_NotionalAvailable": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/governor.NotionalAvailable"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/response.ResponsePagination"
                }
            }
        },
        "response.Response-array_governor_NotionalAvailableDetail": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/governor.NotionalAvailableDetail"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/response.ResponsePagination"
                }
            }
        },
        "response.Response-array_governor_NotionalLimitDetail": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/governor.NotionalLimitDetail"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/response.ResponsePagination"
                }
            }
        },
...
        "response.Response-governor_GovConfig": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/governor.GovConfig"
                },
                "pagination": {
                    "$ref": "#/definitions/response.ResponsePagination"
                }
            }
        },
        "response.Response-governor_GovStatus": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/governor.GovStatus"
                },
                "pagination": {
                    "$ref": "#/definitions/response.ResponsePagination"
                }
            }
        },
        "response.Response-governor_MaxNotionalAvailableRecord": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/governor.MaxNotionalAvailableRecord"
                },
                "pagination": {
                    "$ref": "#/definitions/response.ResponsePagination"
                }
            }
        },
        "response.ResponsePagination": {
            "type": "object",
            "properties": {
                "next": {
                    "type": "string"
                }
            }
        },
...
    }
}